Create session resources
Establish blank `#index`, `#update`, and `#destroy` actions of session
resources. Establish naive authentication (`session_id` as token) and
verify it with specs. `#create` is still route-only.

Note:

During development, there is a modification

    --- /spec/requests/sessions_spec.rb
    +++ /spec/requests/sessions_spec.rb
    @@ -12, +12 @@
    -invalid_session_id = Session.count + 1
    +invalid_session_id = Session.count + 2

Originally `Session.count + 1` is enough to be invalid. However, for
`#update` and `#destroy`, the session is not created until the `Proc` is
called (due to lazy evaluation), which is after line 12 (where
`invalid_session_id` is assigned), but before the request (therefore
before authentication). This makes the "invalid id" valid, so we need to
increment the session ID again.

This will not be a problem with a serious token mechanism, when the
"invalid id" is an unsigned (or tampered) JWT.