Migrate API documentation markup to Markdown
Because in RDoc the `@example_request` and `@example_response` cannot be
properly expressed. In this case both tag only contains a code field;
with RDoc we would write

    @example_response
        !!!json
        {
          "token": "foo"
        }

but the parser don't know that the code block is indented twice; it
regards it the same as

    @example_response
      !!!json
      {
        "token": "foo"
      }

which is parsed as text. We need to tell it that we indent with two
spaces:

    @example_response
      JSON format:

        !!!json
        {
          "token": "foo"
        }

Then it knows that the JSON block is indented twice. However, this
results in duplication of information (which would basically happen
extensively in this API), so I decided to migrate to Markdown (or GitHub
Flavored Markdown).