Requires password only on create
The existance of password was checked everytime the `User` is either
created or updated. Now it is only checked when it is created, since a
user should be allowed to update information without updating the user's
password. This is only a problem for password field, because it is the
only "virtual" field that does not exist in database. Typically
validation should (seem to) happen after the model is updated, so it
will work as expected whether the field is updated or not. Virtual field
like "password" is the only exception.

Note that length validation is still checked on both create and update,
since it is desired to enforce this rule thoughout the database.