Response with JSON array in `#index` of sessions
As explained in `API.markdown`, it should be save to work with array as
top-level JSON. The patch on `Array` class is consistent with the
behavior of the method on `Hash` (added by Rails) acting on a hash of
arrays. For example,

    {
      'a' => [
        {'key1' => 1},
        {'key2' => 2}
      ]
    }

would be transformed by `Hash#with_indifferent_access` into (equivalence
of)

    {
      a: [
        {key1: 1},
        {key2: 2}
      ]
    }

so the same method on Array should (and now does) transform

    [
      {'key1' => 1},
      {'key2' => 2}
    ]

into

    [
      {key1: 1},
      {key2: 2}
    ]

Meta-information was beside the data; now they are (if any) in the
response header.