Support deploying with Docker
Docker officially provides Ruby images based either on `buildpack-deps`,
or on Alpine Linux. The `buildpack-deps` comes with many tools, such as
PostgreSQL, C-compiler toolchain, Git, and Bash, which greatly increased
the size. Alpine Linux, in contrast, forces users to explicitly list
their dependencies, and (in return) gives a slimmer image.

The `bin/entrypoint` is a simple script to wait for database. Better
options exist, such as [this script][1].

[1]: https://github.com/vovimayhem/docker-compose-rails-dev-example/blob/master/bin/check_or_setup_db

New entry in `.gitignore` is commented because `.env` file is also used
in [`dotenv`][2]. If `dotenv` is added in the future, something like
`.env.test` or `.env.development` can be used instead.

[2]: https://github.com/bkeepers/dotenv

Compose file version 2.1 is used instead of version 3 because Xenial
Xerus officially provides Docker 1.12 which doesn't support Compose file
version 3. In addition, swarm mode is not very attractive in foreseeable
future.